requirements.txt
PyQt6
Pillow
rembg[cpu]
opencv-python
numpy
mediapipe
